+++ src/fastmcp/server/context.py
@@ -18,6 +18,7 @@ from pydantic.networks import AnyUrl
 from starlette.requests import Request
 
 from fastmcp.server.server import FastMCP
+from fastmcp.utilities.http import get_current_starlette_request
 from fastmcp.utilities.logging import get_logger
 
 logger = get_logger(__name__)
@@ -59,7 +60,6 @@ class Context(BaseModel, Generic[ServerSessionT, LifespanContextT]):
 
     _request_context: RequestContext[ServerSessionT, LifespanContextT] | None
     _fastmcp: FastMCP | None
-    _request: Request | None
 
     model_config = ConfigDict(arbitrary_types_allowed=True)
 
@@ -68,13 +68,11 @@ class Context(BaseModel, Generic[ServerSessionT, LifespanContextT]):
         *,
         request_context: RequestContext[ServerSessionT, LifespanContextT] | None = None,
         fastmcp: FastMCP | None = None,
-        request: Request | None = None,
         **kwargs: Any,
     ):
         super().__init__(**kwargs)
         self._request_context = request_context
         self._fastmcp = fastmcp
-        self._request = request
 
     @property
     def fastmcp(self) -> FastMCP:
@@ -89,13 +87,6 @@ class Context(BaseModel, Generic[ServerSessionT, LifespanContextT]):
         if self._request_context is None:
             raise ValueError("Context is not available outside of a request")
         return self._request_context
-
-    @property
-    def request(self) -> Request:
-        """Access to the underlying request."""
-        if self._request is None:
-            raise ValueError("Context is not available outside of a request")
-        return self._request
 
     async def report_progress(
         self, progress: float, total: float | None = None
@@ -235,3 +226,10 @@ class Context(BaseModel, Generic[ServerSessionT, LifespanContextT]):
         )
 
         return result.content
+
+    def get_starlette_request(self) -> Request:
+        """Get the active starlette request."""
+        request = get_current_starlette_request()
+        if request is None:
+            raise ValueError("Request is not available outside a Starlette request")
+        return request
